requirements.txt
pymorphy2==0.8
gensim==3.7.3
python-telegram-bot==7.0.1